refactor: split eeBUS proof validator stages
